monitor_app : models refactoring
