+++ monitor_app/models_collection/__pycache__/.gitignore
@@ -1,0 +1,1 @@
+*.pyc
